Add 'popcntgo' tag to force popcnt implementation to be pure-Go
The popcnt instruction isn't available on all amd64 machines, so this is
useful for compatibility purposes.

A future release might check for the available of this instruction at
process init() and behave accordingly.  However, the difference between
the pure-Go version and the assembly version is small enough that any
extra checks in the assembly version might outweigh the performance
gains.  This is perhaps an argument for only having the pure-Go version
in the first place.

Fixes #1